feat(category): Add filtering and pagination for categories

+++ Src/EndPoints/ShahanStore.API/Controllers/CategoryController.cs
@@ -1,5 +1,4 @@
 ﻿using Common.Application.Models.Results;
-using Common.Domain.ValueObjects;
 using MapsterMapper;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
@@ -9,25 +8,25 @@ using ShahanStore.API.Common.FileUtil.Interfaces;
 using ShahanStore.API.DTOs.Categories.Commands;
 using ShahanStore.Application.CQRS.Categories.Commands.AddChild;
 using ShahanStore.Application.CQRS.Categories.Commands.ChangeBanner;
+using ShahanStore.Application.CQRS.Categories.Commands.ChangeIcon;
 using ShahanStore.Application.CQRS.Categories.Commands.Create;
+using ShahanStore.Application.CQRS.Categories.Commands.Delete;
 using ShahanStore.Application.CQRS.Categories.Commands.Edit;
 using ShahanStore.Application.CQRS.Categories.DTOs.Queries;
 using ShahanStore.Application.CQRS.Categories.Queries.GetAll;
+using ShahanStore.Application.CQRS.Categories.Queries.GetByFilter;
 using ShahanStore.Application.CQRS.Categories.Queries.GetById;
-using System.Threading;
-using ShahanStore.Application.CQRS.Categories.Commands.ChangeIcon;
-using ShahanStore.Application.CQRS.Categories.Commands.Delete;
 
 namespace ShahanStore.API.Controllers;
 
 public class CategoryController(IMediator mediator, IMapper mapper, ILocalFileService localFileService) : ApiController
 {
     #region Commands
+
     [HttpPost]
     [ProducesDefaultResponseType(typeof(OperationResult))]
     public async Task<IActionResult> Create([FromForm] CreateCategoryDto request, CancellationToken cancellationToken)
     {
-
         string? bannerImgName = null;
         if (request.BannerImg is not null)
         {
@@ -36,6 +35,7 @@ public class CategoryController(IMediator mediator, IMapper mapper, ILocalFileSe
 
             bannerImgName = await localFileService.SaveFileAsync(request.BannerImg, AppDirectories.CategoryBanner);
         }
+
         string? iconImgName = null;
         if (request.Icon is not null)
         {
@@ -46,7 +46,7 @@ public class CategoryController(IMediator mediator, IMapper mapper, ILocalFileSe
         }
 
 
-        var command =mapper.Map<CreateCategoryCommand>(request);
+        var command = mapper.Map<CreateCategoryCommand>(request);
         var finalCommand = command with { BannerImg = bannerImgName, Icon = iconImgName };
         //var command = new CreateCategoryCommand(
         //    request.Title,
@@ -88,7 +88,8 @@ public class CategoryController(IMediator mediator, IMapper mapper, ILocalFileSe
     [HttpPost]
     [Route("AddChild")]
     [ProducesDefaultResponseType(typeof(OperationResult))]
-    public async Task<IActionResult> AddChild([FromForm] AddChildCategoryDto request, CancellationToken cancellationToken)
+    public async Task<IActionResult> AddChild([FromForm] AddChildCategoryDto request,
+        CancellationToken cancellationToken)
     {
         string? bannerImgName = null;
         if (request.BannerImg is not null)
@@ -98,6 +99,7 @@ public class CategoryController(IMediator mediator, IMapper mapper, ILocalFileSe
 
             bannerImgName = await localFileService.SaveFileAsync(request.BannerImg, AppDirectories.CategoryBanner);
         }
+
         string? iconImgName = null;
         if (request.Icon is not null)
         {
@@ -117,22 +119,20 @@ public class CategoryController(IMediator mediator, IMapper mapper, ILocalFileSe
     [HttpPut]
     [Route("ChangeBanner")]
     [ProducesDefaultResponseType(typeof(OperationResult<string?>))]
-    public async Task<IActionResult> ChangeBanner([FromForm] ChangeCategoryBannerDto request, CancellationToken cancellationToken)
+    public async Task<IActionResult> ChangeBanner([FromForm] ChangeCategoryBannerDto request,
+        CancellationToken cancellationToken)
     {
-        if (request.BannerImg is null&&!request.BannerImg.IsValidImageFile())
+        if (request.BannerImg is null && !request.BannerImg.IsValidImageFile())
             return HandleResult(OperationResult.Error("فایل بنر نامعتبر است"));
 
-        string bannerImgName = await localFileService.SaveFileAsync(request.BannerImg, AppDirectories.CategoryBanner);
+        var bannerImgName = await localFileService.SaveFileAsync(request.BannerImg, AppDirectories.CategoryBanner);
 
         var command = new ChangeCategoryBannerCommand(
             request.CategoryId,
             bannerImgName);
 
         var result = await mediator.Send(command, cancellationToken);
-        if (result.Data is not null)
-        {
-            localFileService.DeleteFile(result.Data, AppDirectories.CategoryBanner);
-        }
+        if (result.Data is not null) localFileService.DeleteFile(result.Data, AppDirectories.CategoryBanner);
         return HandleResult(result);
     }
 
@@ -140,27 +140,38 @@ public class CategoryController(IMediator mediator, IMapper mapper, ILocalFileSe
     [HttpPut]
     [Route("ChangeIcon")]
     [ProducesDefaultResponseType(typeof(OperationResult<string?>))]
-    public async Task<IActionResult> ChangeIcon([FromForm] ChangeCategoryIconDto request, CancellationToken cancellationToken)
+    public async Task<IActionResult> ChangeIcon([FromForm] ChangeCategoryIconDto request,
+        CancellationToken cancellationToken)
     {
         if (!request.Icon.IsValidImageFile() && request.Icon is null)
             return HandleResult(OperationResult.Error("فایل بنر نامعتبر است"));
 
-        string iconImgName = await localFileService.SaveFileAsync(request.Icon, AppDirectories.CategoryIcon);
+        var iconImgName = await localFileService.SaveFileAsync(request.Icon, AppDirectories.CategoryIcon);
 
         var command = new ChangeCategoryIconCommand(
             request.CategoryId,
             iconImgName);
 
         var result = await mediator.Send(command, cancellationToken);
-        if (result.Data is not null)
-        {
-            localFileService.DeleteFile(result.Data, AppDirectories.CategoryIcon);
-        }
+        if (result.Data is not null) localFileService.DeleteFile(result.Data, AppDirectories.CategoryIcon);
         return HandleResult(result);
     }
+
     #endregion
 
     #region Queries
+
+    [HttpGet]
+    [Route("Filter")]
+    [ProducesDefaultResponseType(typeof(OperationResult<CategoryDto>))]
+    public async Task<IActionResult> GetByFilter([FromQuery] CategoryFilterParams filterParams,
+        CancellationToken cancellationToken)
+    {
+        var query = new GetCategoriesByFilterQuery(filterParams);
+        var result = await mediator.Send(query, cancellationToken);
+        return HandleResult(OperationResult<CategoryFilterResult>.Success(result));
+    }
+
 
     [HttpGet]
     [Route("{id:guid}")]
@@ -182,5 +193,6 @@ public class CategoryController(IMediator mediator, IMapper mapper, ILocalFileSe
         var result = await mediator.Send(query, cancellationToken);
         return HandleResult(OperationResult<List<CategoryDto>>.Success(result));
     }
+
     #endregion
 }
